requirements.txt
selene==2.0.0rc9
pytest
allure-pytest
selenium~=4.33.0
python-dotenv